r16293: Fix more format string warnings on a 64-bit box. Jeremy.

+++ source/tdb/io.c
@@ -147,7 +147,7 @@ static int tdb_read(struct tdb_context *tdb, tdb_off_t off, void *buf,
 			/* Ensure ecode is set for log fn. */
 			tdb->ecode = TDB_ERR_IO;
 			TDB_LOG((tdb, 0,"tdb_read failed at %d len=%d ret=%d (%s) map_size=%d\n",
-				 off, len, ret, strerror(errno), tdb->map_size));
+				 off, len, (int)ret, strerror(errno), (int)tdb->map_size));
 			return TDB_ERRCODE(TDB_ERR_IO, -1);
 		}
 	}
